src/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
requests==2.31.0
SQLAlchemy==2.0.25
python-dotenv==1.0.1
apscheduler==3.10.4
openai==1.12.0
solders==0.21.0
pandas==2.2.0
pandas-ta==0.3.14b0
